Update AutoNotify.cs
Getting a lot of warnings in Framework:

Non-nullable property 'PropertyName' must contain a non-null value when exiting constructor. Consider declaring the property as nullable.

This allows for it to be blank to avoid the error or allows the dev to set the PropertyName in the Attribute. I know you do not like allowing to be null:

public string? PropertyName { get; set; }